Workaround to fix VPN build
A previous commit introduces a line of code which requires Tor to be built, which isn't the case for VPN builds.
So script was modified to only have that said line of code called "externally", inside a case called ln_Tor, and rootprep script is called again in the Makefile but this time specifying ln_Tor as a parameter whenever TOR is included.